extract PageLinks and show on 404
