+++ src/app/pages/nav/Layout.tsx
@@ -1,5 +1,6 @@
 import { LayoutProps } from "rwsdk/router";
 import { ToggleNavigationModes } from "@/app/components/ToggleNavigationModes";
+import { PageLinks } from "@/app/components/PageLinks";
 
 export const ViewTransitionLayout = ({ children }: LayoutProps) => {
   return (
@@ -22,18 +23,7 @@ export const ViewTransitionLayout = ({ children }: LayoutProps) => {
       </div>
 
       <br />
-      <div className="flex flex-row gap-2 justify-center">
-        <a href="/nav/page-1" className="text-blue-500">
-          Page 1
-        </a>
-        <a href="/nav/page-2" className="text-blue-500">
-          Page 2
-        </a>
-        <a href="/nav/page-3" className="text-blue-500">
-          Page 3 (404)
-        </a>
-      </div>
-
+      <PageLinks />
       {children}
     </div>
   );
